finalisation_de_la_commande_setlogs

+++ Commands/Configuration/setlogs.js
@@ -10,7 +10,7 @@ module.exports = {
     description: "Je te montre comment gérer les threads !",
     options: [
         {
-            name: 'members',
+            name: 'membres',
             description: 'Salon où sont conservés les logs des arrivées et départs',
             type: ApplicationCommandOptionType.Subcommand,
             options: [
@@ -40,7 +40,7 @@ module.exports = {
             ]
         },
         {
-            name: 'moderation',
+            name: 'modération',
             description: 'Salon où sont conservés les logs des actes de modération',
             type: ApplicationCommandOptionType.Subcommand,
             options: [
@@ -68,6 +68,36 @@ module.exports = {
                     required: false
                 }
             ]
+        },
+        {
+            name: 'autres',
+            description: 'Salon où sont conservés les autres types de logs',
+            type: ApplicationCommandOptionType.Subcommand,
+            options: [
+                {
+                    name: 'choix',
+                    description: `Ajouter ou Supprimer un salon`,
+                    type: ApplicationCommandOptionType.String,
+                    required: true,
+                    choices: [
+                        {
+                            name: 'Ajouter',
+                            value: 'add'
+                        },
+                        {
+                            name: 'Supprimer',
+                            value: 'remove'
+                        },
+                    ]
+                },
+                {
+                    name: 'salon',
+                    description: 'Le salon pour afficher les autres types de logs',
+                    type: ApplicationCommandOptionType.Channel,
+                    channelTypes: 0,
+                    required: false
+                }
+            ]
         }
     ],
     async runInteraction(client, interaction) {
@@ -90,8 +120,8 @@ module.exports = {
         if (!channelTarget) channelTarget = interaction.channel;
         const evtChoices = interaction.options.getString('choix');
 
-        //////////////////////////////// MEMBERS ////////////////////////////////
-        if (interaction.options.getSubcommand() === 'members') {
+        //////////////////////////////// MEMBRES ////////////////////////////////
+        if (interaction.options.getSubcommand() === 'membres') {
             if (evtChoices == 'add') {
                 db.query(`SELECT * FROM logs WHERE type = "members" AND guildID = ${interaction.guild.id}`, async (err, req) => {
                     // S'il n'y a aucun salon enregistré
@@ -165,8 +195,8 @@ module.exports = {
                 })
             }
         }
-        //////////////////////////////// MODERATION ////////////////////////////////
-        else if (interaction.options.getSubcommand() === 'moderation') {
+        //////////////////////////////// MODÉRATION ////////////////////////////////
+        else if (interaction.options.getSubcommand() === 'modération') {
             if (evtChoices == 'add') {
                 db.query(`SELECT * FROM logs WHERE type = "modo" AND guildID = ${interaction.guild.id}`, async (err, req) => {
                     // S'il n'y a aucun salon enregistré
@@ -200,7 +230,7 @@ module.exports = {
                         sucessEmbed.addFields(
                             {
                                 name: `Ok mon pote, j'ai fait la mise à jour !`,
-                                value: `J'ai bien **enregistré** ${channelTarget} comme nouveau salon pour **les les actes de modération** ! Ils s'afficheront là-bas désormais !`
+                                value: `J'ai bien **enregistré** ${channelTarget} comme nouveau salon pour **les actes de modération** ! Ils s'afficheront là-bas désormais !`
                             }
                         )
                         db.query(`UPDATE logs SET channelID = '${channelTarget.id}' WHERE type = "modo" AND guildID = ${interaction.guild.id}`)
@@ -229,6 +259,81 @@ module.exports = {
                             {
                                 name: `Ok mon pote, c'est tout bon !`,
                                 value: `J'ai bien **supprimé** le salon où **les logs des actes de modération** s'affichaient de ma base de données ! Ils ne s'y afficheront plus !`
+                            }
+                        )
+                        let sql = `DELETE FROM logs WHERE type = 'members' AND guildID = ${interaction.guild.id}`
+                        db.query(sql, function(err) {
+                            if(err) throw err;
+                        })
+                        await interaction.reply({ embeds: [sucessEmbed], files: [thumbnailSucess]})
+                    }
+                })
+            }
+        }
+        //////////////////////////////// MODÉRATION ////////////////////////////////
+        else if (interaction.options.getSubcommand() === 'autres') {
+            if (evtChoices == 'add') {
+                db.query(`SELECT * FROM logs WHERE type = "other" AND guildID = ${interaction.guild.id}`, async (err, req) => {
+                    // S'il n'y a aucun salon enregistré
+                    if (req.length < 1) {
+                        sucessEmbed.addFields(
+                            {
+                                name: `Ok mon pote, c'est tout bon !`,
+                                value: `J'ai bien **enregistré** ${channelTarget} dans ma base de données ! **Les logs hors membres et modération** s'afficheront là-bas désormais !`
+                            }
+                        )
+                        let sql = `INSERT INTO logs (type, guildID, channelID) VALUES ('other', '${interaction.guild.id}', '${channelTarget.id}')`
+                        db.query(sql, function (err) {
+                            if (err) throw err;
+                        })
+                        await interaction.reply({ embeds: [sucessEmbed], files: [thumbnailSucess] })
+                    }
+                    // Si le salon est déjà enregistré
+                    else if (req[0].channelID === channelTarget.id) {
+                        let thumbnail = new AttachmentBuilder(`./Img/emotes/grodou2.png`, { name: `grodou.png` })
+                        errorEmbed.setThumbnail(`attachment://${thumbnail.name}`)
+                        errorEmbed.addFields(
+                            {
+                                name: `Oups ! Houston, nous avons un problème !`,
+                                value: `Je viens de me rappeler que ${channelTarget} est déjà enregistré dans la base de données pour **les logs hors membres et modération** !`
+                            }
+                        )
+                        await interaction.reply({ embeds: [errorEmbed], files: [thumbnail] })
+                    }
+                    // Si c'est un autre salon
+                    else {
+                        sucessEmbed.addFields(
+                            {
+                                name: `Ok mon pote, j'ai fait la mise à jour !`,
+                                value: `J'ai bien **enregistré** ${channelTarget} comme nouveau salon pour **les logs hors membres et modération** ! Ils s'afficheront là-bas désormais !`
+                            }
+                        )
+                        db.query(`UPDATE logs SET channelID = '${channelTarget.id}' WHERE type = "other" AND guildID = ${interaction.guild.id}`)
+                        await interaction.reply({ embeds: [sucessEmbed], files: [thumbnailSucess] })
+                    }
+                })
+            }
+            else {
+                db.query(`SELECT * FROM logs WHERE type = "other" AND guildID = ${interaction.guild.id}`, async (err, req) => {
+                    // S'il n'y a aucun salon enregistré
+                    if (req.length < 1) {
+                        let thumbnail = new AttachmentBuilder(`./Img/emotes/grodou2.png`, { name: `grodou.png` })
+                        errorEmbed.setThumbnail(`attachment://${thumbnail.name}`)
+                        errorEmbed.addFields(
+                            {
+                                name: `Oups ! Houston, nous avons un problème !`,
+                                value: `Je peux t'affirmer qu'**aucun salon n'est enregistré** dans ma base de données pour **les logs hors membres et modération** ! De ce fait, impossible d'accéder à ta requête !`
+                            }
+                        )
+
+                        await interaction.reply({ embeds: [errorEmbed], files: [thumbnail] })
+                    }
+                    // Si un salon est déjà enregistré
+                    else {
+                        sucessEmbed.addFields(
+                            {
+                                name: `Ok mon pote, c'est tout bon !`,
+                                value: `J'ai bien **supprimé** le salon où **les logs hors membres et modération** s'affichaient de ma base de données ! Ils ne s'y afficheront plus !`
                             }
                         )
                         let sql = `DELETE FROM logs WHERE type = 'members' AND guildID = ${interaction.guild.id}`
